--- src/App.css
@@ -1,3 +1,0 @@
-.font-inter {
-  font-family: 'inter', sans-serif;
-}
