+++ src/pages/Home/styles.css
@@ -1,0 +1,4 @@
+.font-inter {
+    font-family: 'inter', sans-serif;
+  }
+
